+++ CHANGES.md
@@ -7,11 +7,14 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 ## CHANGE LOG ##
 
 * @dev
+   * **BACKWARD INCOMPATIBLE CHANGES** ([more info](docs/compatibility.md))
    * Created new library logo (see `artwork/` folder)
    * Added more unit tests to improve coverage.
    * Updated documentation.
    * Worked arround Laravel's config merger wnot working properly with multi-dimensional config arrays.
    * Corrected ApiCodesTests trait failing on some methods.
+   * Included ApiCodesTest trait in base tests to avoid desync in future releases.
+   * Removed custom response keys mapping feature.
 
 * v6.0.0 (2019-09-20)
    * **BACKWARD INCOMPATIBLE CHANGES** ([more info](docs/compatibility.md))